@@ -1,6 +1,7 @@
 import shlex
 import subprocess
-from typing import Sequence, Union
+from collections.abc import Sequence
+from typing import Union
 
 import psutil
 
